PLL database service implemented plus tests

+++ app/src/main/java/com/example/smartcubeapp/solvedatabase/dataclasses/OLLData.kt
@@ -6,6 +6,6 @@ data class OLLData(
     val moveCount: Long,
     val startStateID: Long,
     val endStateID: Long,
-    val case: Long,
+    val case: Int,
     val id: Long = -1
 )
